robot/Raspberry Pi/examples/voiceToText/requirements.txt
requests
websocket-client
pyaudio
